More fixes for functionality

+++ routes/user.js
@@ -67,9 +67,21 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-// delete user
+// delete user - will fail if the user has tasks assigned
 router.delete("/:id", async (req, res) => {
   try {
+    // find any tasks assigned to the user
+    const assignedTasks = await Task.find({ assignedTo: req.params.id });
+    if (assignedTasks.length > 0)
+      return res.status(400).send("User has assigned tasks - reassign them");
+    // find any tasks created by the user
+    const createdTasks = await Task.find({ createdBy: req.params.id });
+    if (createdTasks.length > 0)
+      return res
+        .status(400)
+        .send(
+          "User has created tasks - edit them so they are assigned to someone else"
+        );
     const deletedUser = await User.findByIdAndDelete(req.params.id);
     if (!deletedUser) return res.status(404).send("User not found");
     res.status(200).json(deletedUser);
